+++ src/main/kotlin/dev/concert/interfaces/event/ReservationEventListener.kt
@@ -1,11 +1,9 @@
 package dev.concert.interfaces.event
 
-import dev.concert.application.data.DataPlatformFacade
 import dev.concert.application.reservation.ReservationFacade
 import dev.concert.domain.event.reservation.ReservationEvent
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
-import org.springframework.kafka.annotation.KafkaListener
 import org.springframework.scheduling.annotation.Async
 import org.springframework.stereotype.Component
 import org.springframework.transaction.event.TransactionPhase
@@ -13,7 +11,6 @@ import org.springframework.transaction.event.TransactionalEventListener
 
 @Component
 class ReservationEventListener (
-    private val dataPlatformFacade: DataPlatformFacade,
     private val reservationFacade: ReservationFacade,
 ) {
 
@@ -56,26 +53,6 @@ class ReservationEventListener (
         }.onFailure { e ->
             log.error("Kafka Message Send Failed!" , e)
             reservationFacade.changeReservationOutBoxStatusSendFail(event.toEntity().reservationId)
-        }
-    }
-
-    /**
-     * 아웃 박스 패턴 적용
-     *
-     * 1. 아웃박스 상태를 SEND_SUCCESS 로 상태변경
-     * 2. 예약 관련 외부 API 호출 (이로직에서는 Slack)
-     */
-    @Async
-    @KafkaListener(topics = ["reservation"], groupId = "concert_group")
-    fun handleExternalApiKafkaEvent(reservationId : String) {
-        log.info("Kafka Event 수신 성공!!")
-        runCatching {
-            reservationFacade.changeReservationOutBoxStatusSendSuccess(reservationId.toLong())
-            dataPlatformFacade.sendReservationData(reservationId.toLong())
-        }.onFailure { ex ->
-            // 아웃박스 패턴으로 발행은 정상적으로 이루어지는게 보장됨
-            // 비즈니스로직 혹은 외부 api 문제일 수 있으므로 예외를 슬랙으로 보내도록 처리
-            log.error("데이터 플랫폼 전송 에러 : ${ex.message}", ex)
         }
     }
 }
